requirements.txt
python-docx>=1.1.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0
